fix(tavily): land the relocation content changes that previous commit missed
The previous commit (f349f409) ran git mv to move library/ai/tavily/ to
library/developer-tools/tavily/ but the follow-up StrReplace edits never
made it into the index — git commit -m (no -a) only picked up the renames
and the two new research markdowns, leaving the package's content
identical to the pre-rename state. Net result on the remote: directory
moved, but .printing-press.json still said "ai-tools", go.mod still
declared the library/ai/tavily module path, and every internal import +
the -ldflags -X reference still pointed at the old import root — so CI
would have hit both the publish-package category mismatch and a broken
go build.

This commit stages the actual content changes:

- .printing-press.json category: ai-tools -> developer-tools (matches
  the new parent directory)
- go.mod module path -> library/developer-tools/tavily
- every internal import in cmd/, internal/cli/, internal/client/,
  internal/mcp/ rewritten to the new module root
- .goreleaser.yaml -ldflags -X reference updated
- .printing-press-pii-polish.json cli_dir updated

Local: go build ./..., go vet ./..., verify_skill.py, and
verify_publish_package.py --base-ref origin/main all clean.
