+++ library/developer-tools/tavily/.printing-press-pii-polish.json
@@ -1,6 +1,6 @@
 {
   "timestamp": "2026-05-12T17:21:26.794862Z",
-  "cli_dir": "/Users/manikantansrinivasan/Desktop/printing-press-library/library/ai/tavily",
+  "cli_dir": "/Users/manikantansrinivasan/Desktop/printing-press-library/library/developer-tools/tavily",
   "findings": null,
   "findings_count_before": 0
 }
