+++ library/developer-tools/tavily/.printing-press.json
@@ -12,7 +12,7 @@
   "spec_format": "internal",
   "spec_checksum": "sha256:67062802b5df12bb4ccffc644bb998599ed493ec999dc78e3a8efec3cf3eb80b",
   "run_id": "20260512-115845",
-  "category": "ai-tools",
+  "category": "developer-tools",
   "description": "Search the web, extract content, crawl sites, and run deep research from the terminal.",
   "mcp_binary": "tavily-pp-mcp",
   "mcp_tool_count": 7,
