+++ library/developer-tools/tavily/internal/cli/doctor.go
@@ -10,8 +10,8 @@ import (
 	"strings"
 
 	"github.com/spf13/cobra"
-	"github.com/mvanhorn/printing-press-library/library/ai/tavily/internal/client"
-	"github.com/mvanhorn/printing-press-library/library/ai/tavily/internal/config"
+	"github.com/mvanhorn/printing-press-library/library/developer-tools/tavily/internal/client"
+	"github.com/mvanhorn/printing-press-library/library/developer-tools/tavily/internal/config"
 )
 
 // looksLikeDoctorInterstitial reports whether the response body matches a known
