+++ library/developer-tools/tavily/internal/cli/novel_batch.go
@@ -11,7 +11,7 @@ import (
 	"strings"
 
 	"github.com/spf13/cobra"
-	"github.com/mvanhorn/printing-press-library/library/ai/tavily/internal/store"
+	"github.com/mvanhorn/printing-press-library/library/developer-tools/tavily/internal/store"
 )
 
 func newBatchPlanCmd(flags *rootFlags) *cobra.Command {
