+++ library/developer-tools/tavily/internal/cli/novel_drift.go
@@ -8,7 +8,7 @@ import (
 	"fmt"
 
 	"github.com/spf13/cobra"
-	"github.com/mvanhorn/printing-press-library/library/ai/tavily/internal/store"
+	"github.com/mvanhorn/printing-press-library/library/developer-tools/tavily/internal/store"
 )
 
 func newDriftDetectCmd(flags *rootFlags) *cobra.Command {
