+++ library/developer-tools/tavily/internal/cli/novel_research_track.go
@@ -9,7 +9,7 @@ import (
 	"time"
 
 	"github.com/spf13/cobra"
-	"github.com/mvanhorn/printing-press-library/library/ai/tavily/internal/store"
+	"github.com/mvanhorn/printing-press-library/library/developer-tools/tavily/internal/store"
 )
 
 func newResearchTrackCmd(flags *rootFlags) *cobra.Command {
