+++ library/developer-tools/tavily/internal/cli/promoted_web.go
@@ -10,7 +10,7 @@ import (
 	"os"
 
 	"github.com/spf13/cobra"
-	"github.com/mvanhorn/printing-press-library/library/ai/tavily/internal/store"
+	"github.com/mvanhorn/printing-press-library/library/developer-tools/tavily/internal/store"
 )
 
 func newWebPromotedCmd(flags *rootFlags) *cobra.Command {
